ToolbarをDivNoteの子コンポーネントに
